requirements.txt
# PyTool 依赖包列表

# 核心依赖
psutil>=5.8.0          # 系统和进程信息
filelock>=3.0.0        # 文件锁
PyYAML>=6.0            # YAML配置文件支持

# 可选依赖
zmq>=0.0.0             # ZeroMQ进程间通信 (可选)
pydumpling>=0.1.0      # 进程转储功能 (可选)

# 说明:
# - psutil: 用于进程管理和系统信息获取
# - filelock: 用于进程锁功能
# - PyYAML: 用于配置文件管理
# - zmq: 用于进程间通信功能 (如果不使用pcom模块可以不安装)
# - pydumpling: 用于异常转储功能 (如果不使用dump模块可以不安装)